src/test/java/SnailTest.java
import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import java.io.*;

import static org.junit.Assert.assertEquals;

public class SnailTest {

    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();

    @Before
    public void setUpStreams() {
        System.setOut(new PrintStream(outContent));
    }

    @After
    public void cleanUpStreams() {
        System.setOut(null);
        System.setIn(null);
    }

    @Test
    public void testSnailSampleInput() throws IOException {
        String firstTestSampleInput = "10\n" +
                "3\n" +
                "2";

        String expectedResult = "8";

        InputStream testCaseInputStream = new ByteArrayInputStream(firstTestSampleInput.getBytes());
        System.setIn(testCaseInputStream);

        Snail.readInputAndCountDays();

        assertEquals(expectedResult, outContent.toString());
    }
}
